Changed the way the tima is incremented
There was a bug where it would check whether to increment the tima
register.  If the bit where we were checking for the carry had 0 in the
prev div and 1 in the cur div it would increment the tima an extra time.

Pass the cpu_instr test
Fails the instr_timing test, but no longer #255, now gives error it for
each individual instruction