Bug Fixes
Sometimes screens aren't refreshed for a while, resulting delayed
inputs, so added a timer that automatically checks the input 120 times a
second

Becuase the division was floor division, sometimes for the height or
width, the screen would stick to one side of the window.  Changed the
pixSize function so the screen always stays in the center.

Possible Future Fix:

The screen only increases in size once the window has increased or
reduced by 140 or 160, might want to implement a new way to calculate
pixel size using floats for more smooth scaling.