Fixed Sound
There was a bug with the period where the highest bit of the period
wasn't being written in.  This resulted in all of the sounds being much
lower than intended.

Sounds still sounds slightly choppy in a few places, but I'll just leave
it as is for now. At least it is working